+++ aws-java-sdk-s3/src/main/java/com/amazonaws/services/s3/transfer/internal/UploadMonitor.java
@@ -124,7 +124,8 @@ public class UploadMonitor implements Callable<Void>, TransferMonitor {
                                                 multipartUploadCallable.getFutures(),
                                                 multipartUploadCallable.getETags(),
                                                 listener,
-                                                this);
+                                                this,
+                                                multipartUploadCallable.getExpectedPartCount());
 
                 resultFuture.setDelegate(threadPool.submit(completeTask));
             } else {