+++ aws-java-sdk-s3/src/main/java/com/amazonaws/services/s3/transfer/internal/UploadPartRequestFactory.java
@@ -142,4 +142,8 @@ public class UploadPartRequestFactory {
         return totalNumberOfParts;
     }
 
+    public long getOptimalPartSize() {
+        return optimalPartSize;
+    }
+
 }